Add files via upload
Everything works! just change the output stuff to output to a file

+++ Sensor.h
@@ -10,11 +10,11 @@ class Sensor
         Sensor();
         // made constructor to take in x and y coordinates.
         //Sets radius to 5 and battery to 5
-        Sensor(int Xcoor,int Ycoor);
-        int GetX() { return m_X; }
-        void SetX(int val) { m_X = val; }
-        int GetY() { return m_Y; }
-        void SetY(int val) { m_Y = val; }
+        Sensor(float Xcoor,float Ycoor);
+        float GetX() { return m_X; }
+        void SetX(float val) { m_X = val; }
+        float GetY() { return m_Y; }
+        void SetY(float val) { m_Y = val; }
         int GetRadius() { return m_Radius; }
         void SetRadius(int val) { m_Radius = val; }
         int GetBattery() { return m_Battery; }
@@ -22,8 +22,8 @@ class Sensor
         void round_done();
 
     private:
-        int m_X;
-        int m_Y;
+        float m_X;
+        float m_Y;
         int m_Radius;
         int m_Battery;
